Fast forward through summer

+++ src/renderer/io.ts
@@ -9,32 +9,23 @@ if (typeof window !== "undefined" && "Worker" in window) {
 }
 // console.log('+++++++++++++++++++++++++++++++++++++++')
 
-import {
-  PDFJSStatic,
-
-} from "pdfjs-dist";
+import { PDFJSStatic } from "pdfjs-dist";
 const pdfjs: PDFJSStatic = _pdfjs as any;
 
-import {
-  flatten,
-  zeroPad,
-  sortBy
-} from "./utils";
+import { flatten, zeroPad, sortBy } from "./utils";
 
 import { histogram, mean, median, deviation } from "d3-array";
 import { createAutoGrabInfo, createGROBIDMetadata } from "./AutoGrab";
 import { featureToggles } from "../store/featureToggle";
 // const FormData = require('form-data');
 import FormData, { getHeaders } from "form-data";
-import console = require("console");
-
 
 interface FileInfo {
   fullFilePath: string;
   fileNameWithExt: string;
 }
 // todo no spaces for pdfdirs
-
+const dupDir = "_duplicates"
 export const listPdfs = (fullPath: string): Promise<FileInfo[]> =>
   new Promise(resolve => {
     glob(fullPath + "/*.pdf", { nodir: true }, (err, files) => {
@@ -51,7 +42,7 @@ export const listPdfs = (fullPath: string): Promise<FileInfo[]> =>
 export const listDirs = (fullpath: string): Promise<string[]> => {
   return new Promise(resolve => {
     glob(fullpath + "/*/", {}, (err, files) => {
-      resolve(files);
+      resolve(files.filter(file => !file.includes(dupDir)));
     });
   });
 };
@@ -76,27 +67,58 @@ export const setupDirFromPdfs = async (pdfRootDir = "") => {
   const pdfPathInfo = await listPdfs(pdfRootDir);
 
   for (let info of pdfPathInfo) {
-    let neededDir = info.fileNameWithExt
-      .replace(/\.pdf/, "")
-      .replace(/\s/g, "-")
-      .replace(/%/g, "-");
-    let dirNameExists;
-    let count = 0;
-    let dataDir;
+    const filePath = path.join(pdfRootDir, info.fileNameWithExt);
+    var data = new Uint8Array(fs.readFileSync(filePath));
+    const pdf = await pdfjs.getDocument({ data });
+    //@ts-ignore
+    const neededDir = pdf.pdfInfo.fingerprint;
+    const dirNameExists = await fs.pathExists(path.join(pdfRootDir, neededDir));
+    if (dirNameExists) {
+      console.log(info.fileNameWithExt + "is duplicate. ");
+      const dupPath = path.join(pdfRootDir, dupDir)
+      try {
+        await fs.ensureDir(dupPath);
+        await fs.move(info.fullFilePath, path.join(dupPath, info.fileNameWithExt));
+      } catch {
+        console.log(info.fileNameWithExt + " already in duplicate dir. ");
+      }
+      
+      continue;
+    }
 
-    do {
-      count++;
-      dirNameExists = await fs.pathExists(path.join(pdfRootDir, neededDir));
-      const postFix = dirNameExists ? "_" + count : "";
-      neededDir = dirNameExists ? neededDir + postFix : neededDir;
+    // let neededFileName =
 
-      if (count > 100) debugger;
-    } while (dirNameExists);
+    // info.fileNameWithExt
+    //   .replace(/\.pdf/, "")
+    //   .replace(/\s/g, "-")
+    //   .replace(/%/g, "-");
+    // let count = 0;
+    // let dataDir;
+    // do {
+    //   count++;
+    //
+    //   const postFix = dirNameExists ? "_" + count : "";
+    //   neededDir = dirNameExists ? neededDir + postFix : neededDir;
+
+    //   if (count > 100) debugger;
+    // } while (dirNameExists);
 
     await fs.ensureDir(path.join(pdfRootDir, neededDir));
     await fs.move(
       info.fullFilePath,
       path.join(pdfRootDir, neededDir, neededDir + ".pdf")
+    );
+
+    
+    await jsonfile.writeFile(path.join(pdfRootDir, neededDir, "pdfInfo.json"), {
+     //@ts-ignore
+      ...pdf.pdfInfo,
+      originalFileName: info.fileNameWithExt
+    });
+
+    await jsonfile.writeFile(
+      path.join(pdfRootDir, neededDir, info.fileNameWithExt + ".json"),
+      { useThisToSearchByFileName: true }
     );
   }
 
@@ -134,8 +156,8 @@ export const preprocessPdfs = (
   for (let dir of pdfDirs) {
     const files = await ls(dir + "/*");
     const [pdfPath] = files.filter(x => x.endsWith(".pdf"));
-    if (!pdfPath){
-      continue
+    if (!pdfPath) {
+      continue;
     }
     let pdf;
     try {
@@ -143,16 +165,24 @@ export const preprocessPdfs = (
       pdf = await pdfjs.getDocument({ data });
       // pdf = await pdfjs.getDocument(pdfPath);
     } catch (err) {
-      console.log(err+pdfPath);
+      console.log(err + pdfPath);
       debugger;
     }
     const allPageNumbers = [...Array(pdf.numPages).keys()].map(x => x + 1);
 
-    await existsElseMake(
-      path.join(dir, "meta.json"),
-      pdf.getMetadata(), // API defined in https://github.com/mozilla/pdf.js/blob/2a9d195a4350d75e01fafb2c19194b7d02d0a0a5/src/display/api.js#L737
-      overwrite
-    );
+    try {
+      await existsElseMake(
+        path.join(dir, "meta.json"),
+        pdf.getMetadata(), // API defined in https://github.com/mozilla/pdf.js/blob/2a9d195a4350d75e01fafb2c19194b7d02d0a0a5/src/display/api.js#L737
+        overwrite
+      );
+    } catch (err) {
+      await existsElseMake(
+        path.join(dir, "meta.json"),
+        {}, // API defined in https://github.com/mozilla/pdf.js/blob/2a9d195a4350d75e01fafb2c19194b7d02d0a0a5/src/display/api.js#L737
+        overwrite
+      );
+    }
 
     await existsElseMake(
       path.join(dir, "outline.json"),
@@ -174,7 +204,7 @@ export const preprocessPdfs = (
       if (!fileExists || overwrite) {
         const page = await pdf.getPage(pageNumber);
         const viewport = page.getViewport(scale);
-        const text = await page.getTextContent();
+        const text = await page.getTextContent({ normalizeWhitespace: true });
 
         const [xMin, yMin, xMax, yMax] = (viewport as any).viewBox;
         const { width, height } = viewport;
@@ -266,7 +296,6 @@ export const processAutoGrab = (
   //   const dir = pdfDirs[0];
   const pdfDirs = await listDirs(pdfRootDir);
   for (let dir of pdfDirs) {
-
     const files = await ls(dir + "/*");
     const [pdfPath] = files.filter(x => x.endsWith(".pdf"));
     let pdf;
@@ -287,9 +316,11 @@ export const processAutoGrab = (
     // and 2) write to "metadataToHighlight.json"
     // if (featureToggles.showAutoGrab) {
 
-    const fileExists = await fs.pathExists(path.join(dir, "metadataToHighlight.json"));
+    const fileExists = await fs.pathExists(
+      path.join(dir, "metadataToHighlight.json")
+    );
     if (fileExists && !overwrite) {
-      break
+      break;
     }
 
     // TODO: change to Promise.all
@@ -299,7 +330,7 @@ export const processAutoGrab = (
       pdf,
       pdfPath,
       true // Now it always overwrites. TODO (Xin) later, change to variable overwrite
-    ).then(result => console.log("createAutoGrabInfo succeed"))
+    ).then(result => console.log("createAutoGrabInfo succeed"));
   }
   // }
   return pdfDirs;
@@ -309,11 +340,11 @@ export const createAutograbJSON = (
   dir: string,
   overwrite = false
 ) => async () => {
-  console.log("inside createAutograbJSON")
+  console.log("inside createAutograbJSON");
   const files = await ls(dir + "/*");
   const [pdfPath] = files.filter(x => x.endsWith(".pdf"));
-  if (!pdfPath){
-    return false
+  if (!pdfPath) {
+    return false;
   }
   let pdf;
   try {
@@ -327,9 +358,11 @@ export const createAutograbJSON = (
 
   let pagesOfText = await loadPageJson(dir, "textToDisplay");
 
-  const fileExists = await fs.pathExists(path.join(dir, "metadataToHighlight.json"));
+  const fileExists = await fs.pathExists(
+    path.join(dir, "metadataToHighlight.json")
+  );
   if (fileExists && !overwrite) {
-    return false
+    return false;
   }
 
   // TODO: change to Promise.all
@@ -339,20 +372,20 @@ export const createAutograbJSON = (
     pdf,
     pdfPath,
     true // Now it always overwrites. TODO (Xin) later, change to variable overwrite
-  ).then(result => console.log("createAutoGrabInfo succeed for " + dir))
+  ).then(result => console.log("createAutoGrabInfo succeed for " + dir));
 
-  return true
+  return true;
 };
 
 export const createGROBIDJSON = (
   dir: string,
   overwrite = false
 ) => async () => {
-  console.log("inside createAutograbJSON")
+  console.log("inside createAutograbJSON");
   const files = await ls(dir + "/*");
   const [pdfPath] = files.filter(x => x.endsWith(".pdf"));
-  if (!pdfPath){
-    return false
+  if (!pdfPath) {
+    return false;
   }
   let pdf;
   try {
@@ -366,9 +399,11 @@ export const createGROBIDJSON = (
 
   let pagesOfText = await loadPageJson(dir, "textToDisplay");
 
-  const fileExists = await fs.pathExists(path.join(dir, "metadataFromGROBID.json"));
+  const fileExists = await fs.pathExists(
+    path.join(dir, "metadataFromGROBID.json")
+  );
   if (fileExists && !overwrite) {
-    return false
+    return false;
   }
 
   // TODO: change to Promise.all
@@ -376,9 +411,9 @@ export const createGROBIDJSON = (
     path.join(dir, "metadataFromGROBID.json"),
     pdfPath,
     true // Now it always overwrites. TODO (Xin) later, change to variable overwrite
-  ).then(result => console.log("metadataFromGROBID succeed for " + dir))
+  ).then(result => console.log("metadataFromGROBID succeed for " + dir));
 
-  return true
+  return true;
 };
 
 export const processGROBID = (
@@ -392,20 +427,21 @@ export const processGROBID = (
   //   const dir = pdfDirs[0];
   const pdfDirs = await listDirs(pdfRootDir);
   for (let dir of pdfDirs) {
-
     const files = await ls(dir + "/*");
     const [pdfPath] = files.filter(x => x.endsWith(".pdf"));
 
-    const fileExists = await fs.pathExists(path.join(dir, "metadataFromGROBID.json"));
+    const fileExists = await fs.pathExists(
+      path.join(dir, "metadataFromGROBID.json")
+    );
     if (fileExists && !overwrite) {
-      break
+      break;
     }
 
     await createGROBIDMetadata(
       path.join(dir, "metadataFromGROBID.json"),
       pdfPath,
       true // Now it always overwrites. TODO (Xin) later, change to variable overwrite
-    ).then(result => console.log("createGROBIDMetadata succeed"))
+    ).then(result => console.log("createGROBIDMetadata succeed"));
   }
   // }
   return pdfDirs;
@@ -464,17 +500,21 @@ export const loadPageJson = async (
   pageNumbersToLoad: number[] = []
 ) => {
   // todo: make sure filePrefix + ".json" gets made when all pages are done
-  await existsElseMake(
-    path.join(dir, filePrefix + ".json"),
-    preprocessPdfs([dir])
-  );
+  try {
+    await existsElseMake(
+      path.join(dir, filePrefix + ".json"),
+      preprocessPdfs([dir])
+    );
+  } catch {
+    debugger;
+  }
   const finalFile = await jsonfile.readFile(
     path.join(dir, filePrefix + ".json")
   );
   const numberOfPages = finalFile.numberOfPages;
   const pageNumbers = checkGetPageNumsToLoad(numberOfPages, pageNumbersToLoad);
 
-  let pages = [];
+  let pages: PageOfText[] = [];
   for (let pageNum of pageNumbers) {
     const pageId = zeroPad(pageNum, 4);
     const page = await jsonfile.readFile(
@@ -640,6 +680,7 @@ export const loadPdfPages = async (
   pageNumbersToLoad: number[] = [],
   scale = 1
 ) => {
+
   // note this way doesn't work with osx+pdfjs+electron
   //          const pdf = await pdfjs.getDocument(path);
   var data = new Uint8Array(fs.readFileSync(path));
